refactor(system.game.index): migrate to Folio-native routing and render hook
