+++ resources/views/pages/system/[system]/games.blade.php
@@ -1,58 +1,28 @@
 <?php
 
-use function Laravel\Folio\{name};
+use App\Models\System;
+use App\Platform\Services\SystemGamesPageService;
+use Illuminate\View\View;
 
+use function Laravel\Folio\{middleware, name, render};
+
+middleware(['can:viewAny,' . App\Models\System::class]);
 name('system.game.index');
+
+render(function (View $view, System $system, SystemGamesPageService $pageService) {
+    return $view->with($pageService->buildViewData(request(), $system));
+});
 
 ?>
 
-@props([
-    'availableCheckboxFilters' => [],
-    'availableRadioFilters' => [],
-    'availableSelectFilters' => [],
-    'availableSorts' => [],
-    'columns' => [],
-    'filterOptions' => [],
-    'gameListConsoles' => [],
-    'games' => [],
-    'shouldAlwaysShowMetaSurface' => true,
-    'sortOrder' => 'title',
-    'system', // System
-    'totalUnfilteredCount' => null, // ?int
-])
-
-@php
-$pageMetaDescription = '';
-$areFiltersPristine = count(request()->query()) === 0;
-
-if ($areFiltersPristine) {
-    if (empty($games)) {
-        $pageMetaDescription = "There are no games with achievements yet for {$system->Name}. Check again soon.";
-    } else {
-        $numGames = count($games);
-        if ($numGames < 100) {
-            $numGames = floor($numGames / 10) * 10; // round down to the nearest tenth
-        } elseif ($numGames < 1000) {
-            $numGames = floor($numGames / 100) * 100; // round down to the nearest hundredth
-        } else {
-            $numGames = floor($numGames / 1000) * 1000; // round down to the nearest thousandth
-        }
-
-        $localizedCount = localized_number($numGames);
-        // WARNING: If you're tweaking this, try to make sure it doesn't exceed 200 characters.
-        $pageMetaDescription = "Explore {$localizedCount}+ {$system->Name} games on RetroAchievements. Our achievements bring a fresh perspective to classic games, letting you track your progress as you beat and master each title.";
-    }
-}
-@endphp
-
 <x-app-layout
-    pageTitle="{{ $system->Name }} Games"
+    pageTitle="{{ $system->name }} Games"
     :pageDescription="$pageMetaDescription"
 >
     <div>
         <x-system-games-page.system-heading
-            :systemId="$system->ID"
-            :systemName="$system->Name"
+            :systemId="$system->id"
+            :systemName="$system->name"
         />
 
         <x-game.game-list
